test: Turn ad-hoc proptests into `proptest`s
Also, clean up various tests and use the existing test infrastructure
more.

changelog: ignore
